perf(menu/players): possibly resolve perf issues
